backend/web/pyproject.toml
[tool.poetry]
name = "nonamed"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "=3.10.13"
black = "^24.3.0"
fastapi = "^0.110.0"
arrow = "^1.3.0"
uvicorn = "^0.28.1"
boto3 = "^1.34.66"
pymysql = "^1.1.0"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.2.1"
python-multipart = "^0.0.9"
pydantic = {version = "^2.6.4", platform = "manylinux2014_aarch64"}
mangum = "0.11"
PyJWT = "^2.3.0"
sqlalchemy = "^2.0.29"
psycopg2-binary = "^2.9.9"
anthropic = "^0.25.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
